Adjust PlayDrawer platform buttons to centered logos only
